# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SiteSort API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new company and admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{projectId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a new document to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentRequest"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{documentId}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{documentId}/acknowledge:
    post:
      operationId: acknowledgeDocument
      tags: [documents]
      summary: Acknowledge a document (sign-off)
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcknowledgeRequest"
      responses:
        "200":
          description: Acknowledgment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /documents/{documentId}/distribute:
    post:
      operationId: distributeDocument
      tags: [distributions]
      summary: Distribute a document to team members
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistributeRequest"
      responses:
        "200":
          description: Distribution created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /documents/{documentId}/distributions:
    get:
      operationId: getDocumentDistributions
      tags: [distributions]
      summary: Get distribution status for a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Distribution records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Distribution"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    RegisterRequest:
      type: object
      properties:
        companyName:
          type: string
        adminName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        companySize:
          type: string
          enum: [1-10, 11-50, 51-200, 201+]
      required: [companyName, adminName, email, password, companySize]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, project_manager, site_worker, subcontractor]
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
      required: [id, companyId, email, name, role, createdAt]
...
    Document:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        uploadedBy:
          type: string
        uploaderName:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [drawing, method_statement, permit, safety, general]
        version:
          type: integer
        fileUrl:
          type: string
        fileSize:
          type: integer
        previousVersionId:
          type: string
          nullable: true
        status:
          type: string
          enum: [current, superseded]
        requiresAcknowledgment:
          type: boolean
        publicAccess:
          type: boolean
        createdAt:
          type: string
          format: date-time
        distributionSummary:
          $ref: "#/components/schemas/DistributionSummary"
      required: [id, projectId, uploadedBy, uploaderName, name, type, version, fileUrl, fileSize, status, requiresAcknowledgment, publicAccess, createdAt, distributionSummary]
    DocumentDetail:
      allOf:
        - $ref: "#/components/schemas/Document"
        - type: object
          properties:
            distributions:
              type: array
              items:
                $ref: "#/components/schemas/Distribution"
          required: [distributions]
    DistributionSummary:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        viewed:
          type: integer
        acknowledged:
          type: integer
      required: [total, pending, viewed, acknowledged]
    UploadDocumentRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [drawing, method_statement, permit, safety, general]
        fileUrl:
          type: string
        fileSize:
          type: integer
        requiresAcknowledgment:
          type: boolean
        publicAccess:
          type: boolean
        distributeToUserIds:
          type: array
          items:
            type: string
      required: [name, type, fileUrl, fileSize]
    AcknowledgeRequest:
      type: object
      properties:
        pin:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required: []
    DistributeRequest:
      type: object
      properties:
        userIds:
          type: array
          items:
            type: string
      required: [userIds]
    Distribution:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        userId:
          type: string
        userName:
          type: string
        userRole:
          type: string
        status:
          type: string
          enum: [pending, viewed, acknowledged]
        distributedAt:
          type: string
          format: date-time
        viewedAt:
          type: string
          format: date-time
          nullable: true
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, documentId, userId, userName, userRole, status, distributedAt]
...