# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SiteSort API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  companyId: string;
  email: string;
  name: string;
  role: UserRole;
  phone?: string | null;
  createdAt: Date;
  lastActiveAt?: Date | null;
}